+++ src/main/java/org/mariadb/jdbc/client/column/UnsignedSmallIntColumn.java
@@ -66,7 +66,7 @@ public class UnsignedSmallIntColumn extends ColumnDefinitionPacket implements Co
   }
 
   public int getColumnType(final Configuration conf) {
-    return Types.INTEGER;
+    return Types.SMALLINT;
   }
 
   public String getColumnTypeName(final Configuration conf) {
